# Conflicts:
#	src/main/java/eRSPG/Email/EmailEvent.java
#	src/main/java/eRSPG/Repository/BudgetImpl.java
#	src/main/java/eRSPG/controller/ProposalController.java

+++ src/main/java/eRSPG/model/form/ManageBudgetForm.java
@@ -3,29 +3,29 @@ package eRSPG.model.form;
 import java.math.BigDecimal;
 
 public class ManageBudgetForm {
-    private int totalBudget;
-    private int remainingBudget;
-    private int budgetYear;
+    private double totalBudget;
+    private double remainingBudget;
+    private double budgetYear;
 
-    public void setTotalBudget(int totalBudget){
+    public void setTotalBudget(double totalBudget){
         this.totalBudget = totalBudget;
     }
 
-    public int getTotalBudget() {
+    public double getTotalBudget() {
         return this.totalBudget;
     }
 
-    public void setRemainingBudget(int remainingBudget) {
+    public void setRemainingBudget(double remainingBudget) {
         this.remainingBudget = remainingBudget;
     }
 
-    public int getRemainingBudget() {
+    public double getRemainingBudget() {
         return this.remainingBudget;
     }
 
     public void setBudgetYear(int budgetYear) { this.budgetYear = budgetYear; }
 
-    public int getBudgetYear() { return this.budgetYear; }
+    public double getBudgetYear() { return this.budgetYear; }
 
 
 }
